requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
urllib3
